selftests/bpf: add test for signed range tracking through BPF_AND
Add test 'errno_zero_retval_test3' that validates the verifier can now
correctly track signed ranges through the "<<63, s>>63, &-CONST" pattern
produced by Clang's DAGCombiner.

This test exercises the specific case of:
  r0 = random()
  r0 <<= 63       ; move bit 0 to sign bit
  r0 s>>= 63      ; sign-extend to [-1, 0]
  r0 &= -13       ; should yield [-13, 0]

Before the previous patch, the verifier would incorrectly widen the range
on the final AND operation, causing this test to fail. With improved
signed range tracking, the verifier now correctly infers [-13, 0], which
is within the allowed [-4095, 0] range for LSM return values.

This test was originally part of Xu Kuohai's series but was omitted due
to the verifier limitation.

Link: https://lore.kernel.org/bpf/CAADnVQJ2bE0cAp8DNh1m6VqphNvWLkq8p=gwyPbbcdopaKcCCA@mail.gmail.com/
Signed-off-by: Xu Kuohai <xukuohai@huawei.com>
Signed-off-by: Shung-Hsi Yu <shung-hsi.yu@suse.com>